listutils, birthday, syllabic_poetry, mydate
Listutils: Functions that Create, modify or test lists. Used to for all of the remaining files
mydate: supports birthday, functions that work with dates provided by users
Birthday: A simulation that provides statistics on the probability of a certain size group
syllabic_poetry: A program that writes a user a certain type of random poem based on the users choice.